fix(ui): remove formik transitive dependency
