+++ apps/www/app/(main)/door/[code]/(private)/parameters/_client/Student_Editor.tsx
@@ -2,7 +2,7 @@
 
 import { TRPC_React } from ":trpc/client";
 import type { Category } from "@1.modules/category.domain";
-import { OptionCategories } from "@1.modules/category.ui/form/SelectCategoryField";
+import { OptionCategories } from "@1.modules/category.ui/form/select";
 import { Student_Schema, type Student } from "@1.modules/profile.domain";
 import { Button } from "@1.ui/react/button";
 import { input, label } from "@1.ui/react/form/atom";
