+++ packages/@1/modules/category/ui/package.json
@@ -24,7 +24,6 @@
     "@1.modules/category.domain": "workspace:*",
     "@1.ui/react": "workspace:*",
     "@tanstack/react-query": "4.36.1",
-    "formik": "2.4.6",
     "react": "18.3.1",
     "react-content-loader": "7.0.2",
     "tailwind-variants": "0.2.1",
@@ -36,6 +35,7 @@
     "@1.config/tailwindcss": "workspace:*",
     "@1.config/typescript": "workspace:*",
     "@types/react": "18.3.3",
+    "@types/react-dom": "18.3.0",
     "prettier": "3.3.3",
     "tailwindcss": "3.4.9",
     "typescript": "5.5.4"
