+++ packages/@1/modules/category/ui/src/form/select.tsx
@@ -1,0 +1,13 @@
+//
+
+import type { Category } from "@1.modules/category.domain";
+
+//
+
+export function OptionCategories({ categories }: { categories: Category[] }) {
+  return categories.map(({ name, id }) => (
+    <option value={String(id)} key={id}>
+      {name}
+    </option>
+  ));
+}
