+++ packages/@1/modules/exchange/ui/package.json
@@ -33,7 +33,6 @@
     "@tanstack/react-query": "4.36.1",
     "constate": "3.3.2",
     "date-fns": "3.6.0",
-    "formik": "2.4.6",
     "react": "18.3.1",
     "react-aria-components": "1.3.1",
     "react-content-loader": "7.0.2",
